% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marinethermohalinewaves-package.R
\docType{package}
\name{marinethermohalinewaves}
\alias{marinethermohalinewaves-package}
\alias{marinethermohalinewaves}
\title{marinethermohalinewaves: Marine Thermohaline Waves (mtw)}
\description{
Serving curated Marine Thermohaline Wave (mthw) data using R programming language.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/marinethermohalinewaves}
}

}
